chore: remove semantic and syntactic folder, clean up unused files.
