+++ main/src/ca/uwaterloo/flix/api/lsp/provider/CompletionProvider.scala
@@ -18,8 +18,6 @@ package ca.uwaterloo.flix.api.lsp.provider
 import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.api.lsp.*
 import ca.uwaterloo.flix.api.lsp.provider.completion.*
-import ca.uwaterloo.flix.api.lsp.provider.completion.semantic.{GetStaticFieldCompleter, InvokeStaticMethodCompleter}
-import ca.uwaterloo.flix.api.lsp.provider.completion.syntactic.{ExprSnippetCompleter, KeywordCompleter}
 import ca.uwaterloo.flix.language.CompilationMessage
 import ca.uwaterloo.flix.language.ast.TypedAst.Root
 import ca.uwaterloo.flix.language.ast.shared.{SyntacticContext, TraitUsageKind}
@@ -33,7 +31,7 @@ import ca.uwaterloo.flix.language.errors.{ParseError, ResolutionError, TypeError
   * https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_completion
   *
   * This list is not displayed to the user as-is, the client always both sorts and filters the list (or at least,
-  * VSCode does). Therefore we always need to provide both filterText (currently copied from label) and sortText.
+  * VSCode does). Therefore, we always need to provide both filterText (currently copied from label) and sortText.
   *
   * Note that we use textEdit rather than insertText to avoid relying on VSCode's tokenisation, so we can ensure that
   * we're consistent with Flix's parser.
