+++ main/src/ca/uwaterloo/flix/api/lsp/provider/completion/GetStaticFieldCompleter.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Magnus Madsen
+ * Copyright 2025 Chenhao Gao
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,17 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ca.uwaterloo.flix.api.lsp.provider.completion.semantic
+package ca.uwaterloo.flix.api.lsp.provider.completion
 
-import ca.uwaterloo.flix.api.lsp.provider.completion.Completion
-import ca.uwaterloo.flix.api.lsp.provider.completion.Completion.MethodCompletion
+import ca.uwaterloo.flix.api.lsp.provider.completion.Completion.FieldCompletion
 import ca.uwaterloo.flix.language.ast.Name
 import ca.uwaterloo.flix.util.JvmUtils
 
-object InvokeStaticMethodCompleter {
+object GetStaticFieldCompleter {
 
   def getCompletions(clazz: Class[?], field: Name.Ident): List[Completion] = {
-    JvmUtils.getStaticMethods(clazz).sortBy(_.getName).map(MethodCompletion(field, _))
+    JvmUtils.getStaticFields(clazz).sortBy(_.getName).map(FieldCompletion(field, _))
   }
 
 }
